+++ lib/kitchen/directions/bake_link_placeholders.rb
@@ -30,6 +30,9 @@ module Kitchen
             # TODO: log a warning!
             puts "warning! could not find a replacement for '[link]' on an element with ID '#{id}'"
           end
+
+          link_class = book.pantry(name: :link_type).get(id)
+          anchor.add_class(link_class) if link_class.present?
         end
       end
     end