+++ spec/kitchen_spec/directions/bake_numbered_table/v2_spec.rb
@@ -114,4 +114,12 @@ RSpec.describe Kitchen::Directions::BakeNumberedTable::V2 do
       end
     end
   end
+
+  context 'when link to table has class' do
+    it 'stores link class' do
+      pantry = table_with_only_caption_title.pantry(name: :link_type)
+      expect(pantry).to receive(:store).with('table-label', { label: 'tId' })
+      described_class.new.bake(table: table_with_only_caption_title, number: 'S', label_class: 'table-label')
+    end
+  end
 end